Convert the build to use meson instead of cmake
This started out adding the cstdint header to devrfkill.cpp, but than
ran into more and more problems with the cmake build not working. The
structure of plasma applet build appeared to have changed, moc was not
working and needed changes to the original CMakeLists.txt files.
Couldn't figure it out, so switched over to meson before head exploded.